modularize matchup processing from bracket
